add URL to tree after visit, add logging, bump to node12, get list of paths from file
