+++ .nvmrc
@@ -1,1 +1,1 @@
-10.16.3
+12.12.0
