+++ bin/index.js
@@ -2,16 +2,27 @@
 
 'use strict'
 
-const run = require('./run')
+const crawler = require('./web-tree-crawler')
 
 const url = process.argv[2]
-const batchSize = +process.env.BATCH_SIZE
-const cookies = process.env.COOKIES
-const headers = process.env.HEADERS
-const outfile = process.env.OUTFILE
-const timeLimit = +process.env.TIME_LIMIT
 
-run(url, { batchSize, cookies, headers, outfile, timeLimit })
-  .then(console.log)
+const { env } = process
+const batchSize = +(env.batchSize || env.b)
+const cookies = env.cookies || env.c
+const headers = env.headers || env.h
+const outFile = env.outFile || env.o
+const pathList = env.pathList || env.p
+const timeLimit = +env.timeLimit || env.t
+const verbose = (env.verbose || env.v || '').trim() === 'true'
+
+crawler(url, {
+  batchSize,
+  cookies,
+  headers,
+  outFile,
+  pathList
+  timeLimit,
+  verbose
+}).then(console.log)
   .catch(err => console.error(err.message) || 1)
   .then(process.exit)
