+++ test/fixtures/path-list.txt
@@ -1,0 +1,3 @@
+foo
+bar
+baz
